requirements.txt
asgiref==3.2.5
Django==3.0.4
django-cors-headers==3.2.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
joblib==0.14.1
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
stop-words==2018.7.23
Whoosh==2.7.4
